+++ src/recensio/plone/upgrades/v2/20251031082612_disable_livesearch/upgrade.py
@@ -1,0 +1,8 @@
+from ftw.upgrade import UpgradeStep
+
+
+class DisableLivesearch(UpgradeStep):
+    """disable-livesearch."""
+
+    def __call__(self):
+        self.install_upgrade_profile()